+++ blender-qmc/blender-qmc/__init__.py
@@ -21,7 +21,7 @@ bl_info = {
     "name"       : "QMC (Quick Material Colors)",
     "description": "Sets the Base Color of a Material Shader",
     "author"     : "Don Schnitzius",
-    "version"    : (1, 6, 0),
+    "version"    : (1, 7, 0),
     "blender"    : (2, 80, 0),
     "location"   : "3D Viewport > Sidebar > MAT > Quick Material Colors",
     "warning"    : "",
@@ -42,18 +42,10 @@ from .color_sets.globals import *
 
 # BOOLEAN FOR PANEL
 class QMC_SETTINGS(bpy.types.PropertyGroup):
-    rename_material_more: bpy.props.BoolProperty(
-        name='',
-        default=False
-    )
-    active_node_more: bpy.props.BoolProperty(
-        name='',
-        default=False
-    )
-    world_color_more: bpy.props.BoolProperty(
-        name='',
-        default=False
-    )
+    active_node_more:     bpy.props.BoolProperty(name='',default=False)
+    rename_material_more: bpy.props.BoolProperty(name='',default=False)
+    diffuse_more:         bpy.props.BoolProperty(name='',default=False)
+    world_color_more:     bpy.props.BoolProperty(name='',default=False)
 
 
 # PARENT PANEL
@@ -67,23 +59,26 @@ class QMCPanel(bpy.types.Panel):
 
     def draw(self, context):
         layout = self.layout
-        more_bool = context.scene.more_bool
         active_bool = context.scene.active_bool
+        more_bool = context.scene.more_bool
+        diffuse_bool = context.scene.diffuse_bool
         world_bool = context.scene.world_bool
 
         srow = layout.row()
         scol = srow.column(align=True)
         scol.scale_y = 1.25
-        scol.prop(more_bool, "rename_material_more")
         scol.prop(active_bool, "active_node_more")
+        scol.prop(more_bool, "rename_material_more")
+        scol.prop(diffuse_bool, "diffuse_more")
         scol.prop(world_bool, "world_color_more")
 
         scol = srow.column(align=True)
         scol.scale_y = 1.25
         scol.scale_x = 3.0
+        scol.label(text="Selected Node Only")
         scol.label(text="Rename Material")
-        scol.label(text="Selected Node")
-        scol.label(text="World Background")
+        scol.label(text="Set Viewport Color")
+        scol.label(text="Set World Background")
 
 
 # IMPORT PANELS
@@ -138,15 +133,17 @@ def register():
         g.c_icons.load(os.path.splitext(entry)[0], os.path.join(icons_dir, entry), "IMAGE")
     for cls in classes:
         bpy.utils.register_class(cls)
-        bpy.types.Scene.more_bool = bpy.props.PointerProperty(type=QMC_SETTINGS)
         bpy.types.Scene.active_bool = bpy.props.PointerProperty(type=QMC_SETTINGS)
+        bpy.types.Scene.more_bool = bpy.props.PointerProperty(type=QMC_SETTINGS)
+        bpy.types.Scene.diffuse_bool = bpy.props.PointerProperty(type=QMC_SETTINGS)
         bpy.types.Scene.world_bool = bpy.props.PointerProperty(type=QMC_SETTINGS)
 
 
 def unregister():
     bpy.utils.previews.remove(g.c_icons)
-    del bpy.types.Scene.more_bool
     del bpy.types.Scene.active_bool
+    del bpy.types.Scene.more_bool
+    del bpy.types.Scene.diffuse_bool
     del bpy.types.Scene.world_bool
     for cls in reversed(classes):
         bpy.utils.unregister_class(cls)